Add test to ensure OC.reversed has no side effect.
One of the bugs I noticed while working on #20 was that in certain
cases, OC.reversed did not clone deeply enough and could end up mutating
in-place rather than returning a modified copy. This test (which fails on
40aec9f but passes on 4a83832) guards against this error.

This commit doesn't really belong on this branch - should have been part
of #20 but I didn't think of it in time. Figured I'd sneak it in here.